move top-level MainView component to use react-redux
