+++ src/components/Home/Banner.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const Banner = ({ appName, token }) => {
+  if (token) {
+    return null;
+  }
+  return (
+    <div className="banner">
+      <div className="container">
+        <h1 className="logo-font">
+          {appName.toLowerCase()}
+        </h1>
+        <p>A place to share your knowledge.</p>
+      </div>
+    </div>
+  );
+};
+
+module.exports = Banner;
