+++ src/components/Home/Tags.js
@@ -1,0 +1,39 @@
+import React from 'react';
+
+const Tags = props => {
+  const tags = props.tags;
+  if (tags) {
+    return (
+      <div className="tag-list">
+        {
+          tags.map(tag => {
+            const handleClick = ev => {
+              ev.preventDefault();
+              store.dispatch({
+                type: 'APPLY_TAG_FILTER',
+                tag: tag,
+                payload: agent.Articles.byTag(tag)
+              });
+            };
+
+            return (
+              <a
+                href=""
+                className="tag-default tag-pill"
+                key={tag}
+                onClick={handleClick}>
+                {tag}
+              </a>
+            );
+          })
+        }
+      </div>
+    );
+  } else {
+    return (
+      <div>Loading Tags...</div>
+    );
+  }
+};
+
+module.exports = Tags;
